requirements.txt
feedparser==6.0.10
rich==13.7.1
PyYAML==6.0.1
APScheduler==3.10.4
requests==2.31.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
